Refactor Prices context
