+++ src/contexts/Prices/index.ts
@@ -1,0 +1,1 @@
+export { Context, default } from './Prices'
